POOL-356 fix deadlock on massive concurrent requests
Happens when blockWhenExhausted is active.
The idleObjects.takeFirst and .pollFirst operations perform a lock.
So if the pool is empty and the MaxTotal is hit we end up having multiple
Threads perform a lock internally via idleObjects.takeFirst().

But if a condition occurs to destroy the object instead of putting it back into the pool
then nobody notifies the Lock Condition.

The proposed solution will simply create a new object in that case and puts
it back into the idle queue IF there are active threads waiting for
some idle objects.